mixed-agent/foundry/agent/requirements.txt
azure-identity>=1.17.1
agent-framework>=0.1.0
openai>=1.51.0
httpx>=0.27.2
pydantic>=2.9.2
